build: fix compiler warnings

+++ tests/src/database/test.cpp
@@ -78,14 +78,14 @@ TEST_F(DatabaseTest, ReadInitialValues)
         //----------------------------------
         // type section
         // all values should be set to 0 (default type)
-        for (int i = 0; i < IO::Buttons::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Buttons::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::button_t::TYPE, i);
         }
 
         // midi message section
         // all values should be set to 0 (default/note)
-        for (int i = 0; i < IO::Buttons::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Buttons::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::button_t::MESSAGE_TYPE, i);
         }
@@ -101,14 +101,14 @@ TEST_F(DatabaseTest, ReadInitialValues)
 
         // midi velocity section
         // all values should be set to 127
-        for (int i = 0; i < IO::Buttons::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Buttons::Collection::size(); i++)
         {
             DB_READ_VERIFY(127, Database::Config::Section::button_t::VALUE, i);
         }
 
         // midi channel section
         // all values should be set to 1
-        for (int i = 0; i < IO::Buttons::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Buttons::Collection::size(); i++)
         {
             DB_READ_VERIFY(1, Database::Config::Section::button_t::CHANNEL, i);
         }
@@ -117,42 +117,42 @@ TEST_F(DatabaseTest, ReadInitialValues)
         //----------------------------------
         // enable section
         // all values should be set to 0
-        for (int i = 0; i < IO::Encoders::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Encoders::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::encoder_t::ENABLE, i);
         }
 
         // invert section
         // all values should be set to 0
-        for (int i = 0; i < IO::Encoders::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Encoders::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::encoder_t::INVERT, i);
         }
 
         // mode section
         // all values should be set to 0
-        for (int i = 0; i < IO::Encoders::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Encoders::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::encoder_t::MODE, i);
         }
 
         // midi id section
         // incremental values - first value should be 0, each successive value should be incremented by 1 for each group
-        for (int i = 0; i < IO::Encoders::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Encoders::Collection::size(); i++)
         {
             DB_READ_VERIFY(i, Database::Config::Section::encoder_t::MIDI_ID, i);
         }
 
         // midi channel section
         // all values should be set to 1
-        for (int i = 0; i < IO::Encoders::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Encoders::Collection::size(); i++)
         {
             DB_READ_VERIFY(1, Database::Config::Section::encoder_t::CHANNEL, i);
         }
 
         // pulses per step section
         // all values should be set to 4
-        for (int i = 0; i < IO::Encoders::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Encoders::Collection::size(); i++)
         {
             DB_READ_VERIFY(4, Database::Config::Section::encoder_t::PULSES_PER_STEP, i);
         }
@@ -161,21 +161,21 @@ TEST_F(DatabaseTest, ReadInitialValues)
         //----------------------------------
         // enable section
         // all values should be set to 0
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::analog_t::ENABLE, i);
         }
 
         // invert section
         // all values should be set to 0
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::analog_t::INVERT, i);
         }
 
         // type section
         // all values should be set to 0
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::analog_t::INVERT, i);
         }
@@ -192,35 +192,35 @@ TEST_F(DatabaseTest, ReadInitialValues)
 
         // lower limit section
         // all values should be set to 0
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::analog_t::LOWER_LIMIT, i);
         }
 
         // upper limit section
         // all values should be set to 16383
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(16383, Database::Config::Section::analog_t::UPPER_LIMIT, i);
         }
 
         // midi channel section
         // all values should be set to 1
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(1, Database::Config::Section::analog_t::CHANNEL, i);
         }
 
         // lower offset section
         // all values should be set to 0
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::analog_t::LOWER_OFFSET, i);
         }
 
         // upper offset section
         // all values should be set to 0
-        for (int i = 0; i < IO::Analog::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Analog::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::analog_t::UPPER_OFFSET, i);
         }
@@ -246,28 +246,28 @@ TEST_F(DatabaseTest, ReadInitialValues)
 
         // rgb enable section
         // all values should be set to 0
-        for (int i = 0; i < IO::LEDs::Collection::size() / 3 + (IO::Touchscreen::Collection::size() / 3); i++)
+        for (size_t i = 0; i < IO::LEDs::Collection::size() / 3 + (IO::Touchscreen::Collection::size() / 3); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::leds_t::RGB_ENABLE, i);
         }
 
         // control type section
         // all values should be set to MIDI_IN_NOTE_MULTI_VAL
-        for (int i = 0; i < IO::LEDs::Collection::size(); i++)
+        for (size_t i = 0; i < IO::LEDs::Collection::size(); i++)
         {
             DB_READ_VERIFY(static_cast<uint32_t>(IO::LEDs::controlType_t::MIDI_IN_NOTE_MULTI_VAL), Database::Config::Section::leds_t::CONTROL_TYPE, i);
         }
 
         // activation value section
         // all values should be set to 127
-        for (int i = 0; i < IO::LEDs::Collection::size(); i++)
+        for (size_t i = 0; i < IO::LEDs::Collection::size(); i++)
         {
             DB_READ_VERIFY(127, Database::Config::Section::leds_t::ACTIVATION_VALUE, i);
         }
 
         // midi channel section
         // all values should be set to 1
-        for (int i = 0; i < IO::LEDs::Collection::size(); i++)
+        for (size_t i = 0; i < IO::LEDs::Collection::size(); i++)
         {
             DB_READ_VERIFY(1, Database::Config::Section::leds_t::CHANNEL, i);
         }
@@ -297,56 +297,56 @@ TEST_F(DatabaseTest, ReadInitialValues)
 
         // x position section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::X_POS, i);
         }
 
         // y position section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::Y_POS, i);
         }
 
         // width section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::WIDTH, i);
         }
 
         // height section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::HEIGHT, i);
         }
 
         // on screen section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::ON_SCREEN, i);
         }
 
         // off screen section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::OFF_SCREEN, i);
         }
 
         // page switch enabled section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::PAGE_SWITCH_ENABLED, i);
         }
 
         // page switch index section
         // all values should be set to 0
-        for (int i = 0; i < IO::Touchscreen::Collection::size(); i++)
+        for (size_t i = 0; i < IO::Touchscreen::Collection::size(); i++)
         {
             DB_READ_VERIFY(0, Database::Config::Section::touchscreen_t::PAGE_SWITCH_INDEX, i);
         }
@@ -445,7 +445,7 @@ TEST_F(DatabaseTest, LEDs)
     ASSERT_TRUE(_database._instance.update(Database::Config::Section::leds_t::CONTROL_TYPE, 0, IO::LEDs::controlType_t::PC_SINGLE_VAL));
 
     // rgb state shouldn't change
-    for (int i = 0; i < IO::LEDs::Collection::size() / 3; i++)
+    for (size_t i = 0; i < IO::LEDs::Collection::size() / 3; i++)
     {
         DB_READ_VERIFY(0, Database::Config::Section::leds_t::RGB_ENABLE, i);
     }
